+++ pyccel/codegen/python_wrapper.py
@@ -6,10 +6,10 @@
 
 import os
 
-from pyccel.ast.bind_c                      import as_static_module
 from pyccel.ast.numpy_wrapper               import get_numpy_max_acceptable_version_file
 from pyccel.codegen.printing.fcode          import fcode
 from pyccel.codegen.printing.cwrappercode   import CWrapperCodePrinter
+from pyccel.codegen.wrapper.fortran_to_c_wrapper   import FortranToCWrapper
 from pyccel.codegen.utilities      import recompile_object
 from pyccel.codegen.utilities      import copy_internal_library
 from pyccel.codegen.utilities      import internal_libs
@@ -36,6 +36,58 @@ def create_shared_library(codegen,
                           wrapper_compiler,
                           sharedlib_modname=None,
                           verbose = False):
+    """
+    Create a shared library which can be called from Pyccel.
+
+    From a CodePrinter object describing code which has been printed
+    in a target language, create a shared library which can be
+    called from Pyccel. In order to do this the code must be wrapped.
+    First, if the code is not written in C, it must be wrapped to
+    make it callable from C. This intermediary code is printed
+    and compiled. From the C-compatible code a second (first for C)
+    wrapper is created which exposes the C code to Python. This
+    is done via the CWrapper. Finally this new code is compiled
+    to generate the required shared language.
+
+    Parameters
+    ----------
+    codegen : pyccel.codegen.printing.codeprinter.CodePrinter
+        The printer which was used to print the translated code.
+
+    main_obj : pyccel.codegen.compiling.basic.CompileObj
+        The compile object which describes the translated code.
+
+    language : str
+        The language which Pyccel translated to.
+
+    wrapper_flags : iterable
+        Any additional flags which should be used to compile the wrapper.
+
+    pyccel_dirpath : str
+        The path to the directory where the files are created and compiled.
+
+    src_compiler : pyccel.codegen.compiling.compilers.Compiler
+        The compiler which should be used to compile the library.
+
+    wrapper_compiler : pyccel.codegen.compiling.compilers.Compiler
+        The compiler which should be used to compile the wrapper.
+        Often this is the same as src_compiler but it may be different
+        when the language is not C to ensure that src_compiler can link
+        the appropriate language-specific libraries.
+
+    sharedlib_modname : str, default: None
+        The name of the shared library. The default is the name of the
+        module printed by the printer.
+
+    verbose : bool, default: False
+        Indicates if the compiling should be done with verbosity to show the
+        compiler commands.
+
+    Returns
+    -------
+    str
+        The absolute path to the shared library which was created.
+    """
 
     pyccel_stage.set_stage('cwrapper')
 
@@ -60,7 +112,8 @@ def create_shared_library(codegen,
 
     if language == 'fortran':
         # Construct static interface for passing array shapes and write it to file bind_c_MOD.f90
-        bind_c_mod = as_static_module(codegen.routines, codegen.ast)
+        wrapper = FortranToCWrapper()
+        bind_c_mod = wrapper.wrap(codegen.ast)
         bind_c_code = fcode(bind_c_mod, bind_c_mod.name)
         bind_c_filename = '{}.f90'.format(bind_c_mod.name)
 
@@ -75,6 +128,9 @@ def create_shared_library(codegen,
         src_compiler.compile_module(compile_obj=bind_c_obj,
                 output_folder=pyccel_dirpath,
                 verbose=verbose)
+        c_ast = bind_c_mod
+    else:
+        c_ast = codegen.ast
 
     #---------------------------------------
     #     Compile cwrapper from stdlib
@@ -100,7 +156,7 @@ def create_shared_library(codegen,
     codegen.ast.set_name(sharedlib_modname)
     wrapper_codegen = CWrapperCodePrinter(codegen.parser.filename, language)
     Scope.name_clash_checker = name_clash_checkers['c']
-    wrapper_code = wrapper_codegen.doprint(codegen.ast)
+    wrapper_code = wrapper_codegen.doprint(c_ast)
     if errors.has_errors():
         return
 
@@ -112,8 +168,8 @@ def create_shared_library(codegen,
     #--------------------------------------------------------
     #  Compile cwrapper_ndarrays from stdlib (if necessary)
     #--------------------------------------------------------
-    if "ndarrays" in wrapper_codegen.get_additional_imports():
-        for lib_name in ("ndarrays", "cwrapper_ndarrays"):
+    for lib_name in ("ndarrays", "cwrapper_ndarrays"):
+        if lib_name in wrapper_codegen.get_additional_imports():
             stdlib_folder, stdlib = internal_libs[lib_name]
 
             lib_dest_path = copy_internal_library(stdlib_folder, pyccel_dirpath)