+++ tests/epyccel/modules/augassign.py
@@ -153,3 +153,14 @@ def augassign_div_2d_complex(a):
     b = a
     b /= (4.0 + 2.0j)
     return b[0][0]
+
+def augassign_func(x : float, y : float):
+    def fun1(x: 'float') -> 'float':
+        return x + 1
+    x %= fun1(y)
+    return x
+
+def augassign_array_func(x : 'float[:]', y : 'float[:]'):
+    def fun1(x: 'float[:]') -> 'float[:]':
+        return x + 1
+    x %= fun1(y)