+++ tests/errors/syntax_errors/empty_class.py
@@ -1,0 +1,3 @@
+# pylint: disable=missing-function-docstring, missing-module-docstring, missing-class-docstring
+class A(object):
+    pass